requirements.txt
appdirs==1.4.4
argh==0.26.2
arrow==1.2.2
astroid==2.11.5
attrs==21.4.0
Babel==2.10.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.3.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cookiecutter==1.7.3
cryptography==37.0.2
css-html-js-minify==2.5.5
dill==0.3.4
dnspython==2.2.1
docutils==0.17.1
elementpath==2.5.1
esbonio==0.11.2
Faker==13.11.0
flake8==4.0.1
future==0.18.2
ghp-import==2.1.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
lazy-object-proxy==1.7.1
livereload==2.6.3
lunr==0.6.2
lxml==4.8.0
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
mergedeep==1.3.4
mkautodoc==0.1.0
mkdocs==1.2.3
mkdocs-autorefs==0.4.1
mkdocs-bootstrap==1.1
mkdocs-material==8.2.7
mkdocs-material-extensions==1.0.3
mkdocs-print-site-plugin==2.3.3
mkdocstrings==0.17.0
mypy==0.950
mypy-extensions==0.4.3
myst-parser==0.17.2
nltk==3.7
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
poyo==0.5.0
psutil==5.9.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.4.0
pygls==0.11.3
Pygments==2.12.0
pylint==2.13.8
pymdown-extensions==9.4
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pyspellchecker==0.6.3
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytkdocs==0.16.1
pytz==2022.1
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.4.24
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.9
watchdog==2.1.7
wrapt==1.14.1
xmlschema==1.10.0
zipp==3.8.0
